+++ crates/shielder-scheduler-server/src/storage/mod.rs
@@ -24,7 +24,7 @@ pub trait StorageProvider: Send + Sync {
 
     /// Get pending requests that are due for processing.
     /// Limits the number of returned requests to `limit`.
-    /// Pending requests are those with status `Pending` or `Processing`
+    /// Pending requests are those with status `Pending`
     /// and `relay_after` time less than or equal to the current time.
     fn get_pending_requests(
         &self,
@@ -49,15 +49,13 @@ pub trait StorageProvider: Send + Sync {
         &self,
         last_note_index: &str,
         new_relay_after: DateTime<Utc>,
-        new_retry_count: i32,
+        new_retry_count: u8,
         processed_at: Option<DateTime<Utc>>,
         new_error_message: Option<&str>,
     ) -> impl std::future::Future<Output = Result<(), StorageError>> + Send;
 
     /// Retrieve a scheduled request by its last note index.
     /// Returns `Ok(None)` if no such request exists.
-    /// If multiple requests have the same last note index,
-    /// returns the one with the earliest `created_at` timestamp.
     fn get_request_by_last_note_index(
         &self,
         last_note_index: &str,
